xetblob reads with mdbv2 (#27)
1. Moved some functions used both by PFTv1 and PFTv2 from
data_processing_v1.rs to data_processing.rs. They were already extracted
out of `PointerFileTranslatorV1` but remained inside
data_processing_v1.rs and PointerFileTranslatorV2 had to reference to
them by data_processing_v1::. This helps keep xetblob code clean.

2. Makes git fetch only merkledb* notes and reposalt note if it exists.
